notebook_test_docs/pyproject.toml
[tool.poetry]
name = "notebook-test-docs"
version = "0.1.0"
description = ""
authors = ["Volodymyr Kurov <vokur13@gmail.com>"]
readme = "README.md"
packages = [{include = "notebook_test_docs"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.103.2"
uvicorn = "^0.23.2"
sqlalchemy = "^2.0.21"
psycopg2-binary = "^2.9.7"
libgravatar = "^1.0.4"
python-jose = "^3.3.0"
passlib = "^1.7.4"
python-multipart = "^0.0.6"
alembic = "^1.12.0"
email-validator = "^2.0.0.post2"
bcrypt = "^4.0.1"
fastapi-mail = "^1.4.1"
redis = "^5.0.0"
python-dotenv = "^1.0.0"
pydantic-settings = "^2.0.3"
slowapi = "^0.1.8"
limiter = "^0.3.1"
cloudinary = "^1.34.0"


[tool.poetry.group.dev.dependencies]
sphinx = "^7.2.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
